+++ lib/on_strum/healthcheck/core.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module OnStrum
-  module GemName
-    VERSION = '0.1.0'
+  module Healthcheck
+    require_relative 'version'
   end
 end